admin/requirements.txt
attr==0.3.2
ConfigParser==7.1.0
contextlib2==21.6.0
cryptography==44.0.0
Cython==3.0.11
dl==0.1.0
docutils==0.21.2
filelock==3.17.0
HTMLParser==0.0.2
ipython==8.12.3
ipywidgets==8.1.5
Jinja2==3.1.5
keyring==25.6.0
mock==5.1.0
numpy==2.2.2
Pillow==11.1.0
protobuf==5.29.3
pyOpenSSL==25.0.0
pywin32==308
railroad==0.5.0
redis==5.2.1
Sphinx==8.1.3
thread==2.0.5
tornado==6.4.2
tqdm==4.67.1
trove_classifiers==2025.1.15.22
urllib3_secure_extra==0.1.0
xmlrpclib==1.0.1
